# docs/swagger.yaml
basePath: /api/v1
definitions:
  project_internal_presentation_schemas.AuthorizeSchema:
    properties:
      password:
        example: 1Q2w3e
        type: string
      username:
        example: maximEZ
        type: string
    type: object
  project_internal_presentation_schemas.CreateUserSchema:
    properties:
      email:
        example: maxim@mail.ru
        type: string
      password:
        example: 1Q2w3e
        type: string
      username:
        example: maximEZ
        type: string
    type: object
  project_internal_presentation_schemas.ErrorSchema:
    properties:
      error:
        type: string
    type: object
  project_internal_presentation_schemas.TokensSchema:
    properties:
      access_token:
        type: string
    type: object
  project_internal_presentation_schemas.UserCreatedSchema:
    properties:
      user_id:
        example: 1
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  title: Go Messenger API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Учётные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project_internal_presentation_schemas.AuthorizeSchema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project_internal_presentation_schemas.TokensSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project_internal_presentation_schemas.ErrorSchema'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/project_internal_presentation_schemas.ErrorSchema'
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project_internal_presentation_schemas.CreateUserSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project_internal_presentation_schemas.UserCreatedSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/project_internal_presentation_schemas.ErrorSchema'
      summary: Регистрация пользователя
      tags:
      - auth
swagger: "2.0"